BUG#35180061 mysqlsh produces wrong dump in case of complex primary key
When dumping with `chunking` enabled, if PK or a unique index selected
to chunk a table contained an ENUM column with values specified in
non-alphabetic order, some rows from that table were missing in the
dump.

The chunking algorithm is using '>' and '<' operators to create ranges
defining each chunk. In case of an ENUM, these operators use
lexicographical ordering of the string literals assigned to the index
values and not the index values themselves. If values of an ENUM column
are not specified in the alphabetic order, chunks will not cover the
whole table and some rows will be missing.

As a fix, PKs and unique indexes which contain one or more ENUM columns
are ignored when selecting an index for the chunking algorithm.

Change-Id: I188351135231d8821c0bb26b6c0b9badc243acf6