Detect if trace plugin was successfully registered
The libmysqlclient supports only one trace client at a time, trying to
register another one always fails. The library can be built with the
WITH_TEST_TRACE_PLUGIN flag enabled, in this case it has a
pre-registered trace plugin. In such case, the event handler plugin used
by Shell fails to register, and the plugin-based authentication does not
work.

Added a check to ensure that the plugin was successfully registered.

Change-Id: Id2dca6a7d019090da975a98c9311917cb9754682