WL#15561 tests postfix - No oci config file fails tests on public tests.
Public build tests don't have defined config file, because of security reasons. The defined env variable OCI_AUTH_CONFIG_FILE has an empty string value for this tests, and that causes them to fail. This commits adds conditions for this tests, to be skipped when OCI_AUTH_CONFIG_FILE is empty.

Change-Id: I096d6036ed5a9f5ed63f4d08c424d6ff2e226c87