BUG#35354749 Incorrect GR transaction_size_limit after calling rejoinCluster
If cluster.rejoinCluster is called with clusters that are ONLINE, it
should behave as a no-op. In other word, in a perfectly healthy
ClusterSet, rejoining a Cluster should not produce an error and should
simply behave as a no-op. However, in that same scenario, if the cluster
is the primary one, the group_replication_transaction_size_limit
variable is updated incorrectly, as if it was a secondary cluster.

This patch fixes this issue by simply not updating that limit when
rejoining the primary cluster.

Change-Id: If2b11bb8650405e5de8713cf44247274bab6fee7