WL#15601 BUG#35211461 AdminAPI: Support stats_updates_frequency option in Cluster.setRoutingOption
WL#15599 adds support for setting the frequency of periodic updates
that Router does in the Metadata on standalone Clusters and ReplicaSets.
On the Shell, WL#15056 added support for setting routing options at the
Cluster level.

This WL builds on both WLs by adding the same kind of support to
ReplicaSets and also adding support for the "stats_updates_frequency"
option for both Clusters and ReplicaSets.

Change-Id: Ifd385bd37dbd0f50d8b5bb58b6bfa08368083347