Updating testenv to work based on a Jenkins server with required authentication.
The logic on the testenv script was based on the fact that no authentication was
required to perform the following operations:

- Job triggering
- Job status retrieval

When a Jenkins server has authentication enabled, Job Triggering can be achieved
using a token specified in the Job, but there's no way to do Job status retrieval.

In addition to the above, traces will now be located at a different server than
the one running CI.

This patch updates the testenv logic to make it based on the availaility of the
traces for a given patch.

Additionally, testing for socket authentication is disabled if the system user
running the tests is root.

Change-Id: If533ff46062ca0da6195a64cd0562d72dc8dc937