BUG#35279351 util.importTable() from a single uncompressed file may fail
The util.importTable() imported single uncompressed files in chunks of
size of `bytesPerChunk`. The chunking algorithm which splits the file in
chunks at row boundaries was looking for an unescaped sequence of
`fieldsTerminatedBy` characters. This approach would produce invalid
chunks (with rows split in the middle of a field) if a field contained
such unescaped sequence, which is allowed by the MySQL server.

As a workaround for this issue, another approach for chunking is used.
If a single file is imported, additional threads are spawned (the number
is controlled by the `threads` option). These threads read the file in
blocks of a fixed size, then these blocks are scanned sequentially by a
single thread, looking for row boundaries using the same algorithm as
implemented by the MySQL server. Once read, the file data is kept in
memory until it is loaded by the server, but the memory consumption is
capped at approximately `bytesPerChunk` * (1 + `threads`) - read
operation is suspended until one of the loading threads finishes its
work.

This approach applies to compressed files, which are now also chunked.

Additional changes:

Disable chunking and sub-chunking if `linesTerminatedBy` is set to an
empty string, or if it's set to the same value as `fieldsTerminatedBy`.

Progress reporting when importing a single compressed file uses the
uncompressed data size as the total size. Progress is displayed as:
  ?% (58.11 MB / ?), 28.15 MB/s
Once the uncompressed data size is known, it switches to normal
reporting:
  86% (671.09 MB / 771.87 MB), 27.26 MB/s

Change-Id: I29890c1582590bcda7acb94738fdd4ada7440f29
(cherry picked from commit 3583427a25c44f2f6e4647e0d9ada6e13a86d889)