BUG#35114864  Improve NOTE: The Cluster's group_replication_view_change_uuid
This patch improves the note shown when rescanning a cluster and the
group_replication_view_change_uuid var is set (persisted) but not yet
being used.

The new note is more explicit regarding the interaction between the var
and if the Cluster is to be used in a ClusterSet. Also, a warning
telling the user to call rebootClusterFromCompleteOutage (to force the
var to be used) was removed, mainly because it is unnecessary: if the
user tries to create a ClusterSet, a check is made and the command fails
if group_replication_view_change_uuid hasn't the correct value.

Change-Id: I76c2427ad7cee300c0270f7a57d5a818f75cbf58