BUG#35077966 \sql does not print statement id despite correct server config and classic
The shell appends 'sql_mode' to the 'session_track_system_variables' variable even when its value is
'*' which is used to enable tracking all the session variables and should not be used with other values.
This patch fixes the problem by only appending 'sql_mode' when 'session_track_system_variables' is
not '*' and 'sql_mode' is missing.

Change-Id: I7da8a2969bee0149a131daae83882e7261e6619e