BUG#35403452 PAR Dump Enhancements: File Buffering and Worker Session Pool
When dumping using PAR, the workers split the data dump operation in 2 parts:

- Buffering data
- Transferring data to Bucket

This patch replaces the memory buffering with file buffering. The dump files will
first be dumped in local disk to later be transferred to the target bucket.
The location of the files will in the system default temporary folder.

Additionally, a new session pool is introduced on the dumper to enable the
workers to release the database session as soon as the buffering step is complete.
This way, the worker thread can continue with the data transfer while another
worker can pick the next dump operation.

Also the number of workers when dumping with PAR is internally doubled, so there
can be up to 'threads'*2 workers doing dump/transfer operations while there are
at most 'threads' workers doing dump operations.

Notes on System Temporary Directory
-----------------------------------

The following order of precedence is used to determine the value of the temporary
directory:

POSIX: TMPDIR, TMP, TEMP, TEMPDIR. If none is defined uses /tmp.

Windows: TMP, TEMP, USERPROFILE. If none defined, uses the Windows directory.

Change-Id: I99c534c6b24857ceb7544a14f847cf5a4c7bb9c9