BUG#35400167 Read replica not shown in cluster.status()
If a Read-Replica enters an ERROR state due to a connection error for
some reason (for example, it's missing transactions that were purged),
and its current source is removed from the Cluster, the Read-Replica is
no longer shown in the output of cluster.status().

This patch fixes that by adding a missing check to verify if the current
read-replica source still belongs to the Cluster or not, to determine
whether the read-replica must be considered rogue and placed under the
current primary.

Change-Id: I504e266017f2af4a62b6afcc4ab223b481ad1e21