Fix for several tests failing on macOS because of mismatch traces
The WL#15056 patch updated code when checking a Cluster recovery method,
mainly calling two methods that perform DB queries. The way the methods
are being called however, can change, causing a mismatch on the order of
the generated traces.

This patch fixes this by forcing an order on the methods.

Change-Id: I32bdcc10341e3d1eb2006b6657674b0f0127f7d2