Post-fix for read_only_targets handling when creating a ClusterSet:
With WL#15056, the handling for read_only_targets made sure that when
creating a ClusterSet from a Cluster the value for read_only_targets is
copied over to the ClusterSet, i.e. the ClusterSet value inherits the
current value of the target Cluster.

However, the value stored in the clusters table was deleted too when it
should not.

This patch removes that removal step.

Change-Id: I66080ed77bca4e22cdf2e05da96fb3e954338152