BUG#35410360 Admin API X Protocol Port changes
The AdminAPI stores in the Metadata schema information about each
instance's endpoints, including the X Plugin port endpoint. This
particular endpoint is used by Router when the GR notifications are
enabled.

If, for some reason, the X Port is changed on a Cluster member and the
member is restarted, the information stored in the Metadata will be
outdated and the AdminAPI is not detecting it or attempting to correct
it. This can lead to MySQL Router not being able to connect for the GR
notifications or connect to a wrong Cluster.

This patch fixes that by including a check for a possible change of the
X Port in Cluster.status() and the possibility of updating the Metadata
schema with Cluster.rescan().

Change-Id: I46c96fd397b1f7dd48c325448adb2f1437385da4