BUG#35358531 ClusterSet.createReplicaCluster proposing non usable clone provisioning
By default, all commands that allow clone-based recovery will pick a
clone donor if not explicitly set using the 'cloneDonor' option.
However, the validations regarding donor and recipient are only
performed when the cloneDonor option is used.

This patch fixes that by ensuring the usage of clone is valid for
recovery and fail soon avoiding any execution of steps and possible
revert of operations.

Other than that, this patch also fixes a bug in rejoinInstance() for
Read-Replicas on which the cloneDonor option was being ignored so a
donor would be picked automatically.

Change-Id: I9f8983c0a2f2a78744de9c3a04cde0873875a811