BUG#35416666 Incorrect SSL options when rebooting a cluster
When rebooting a cluster configured with SSL options (i.e.:
memberAuthType) and switching the comm stack to "mysql", the seed
instance isn't rebooted with the correct SSL options.

This patch fixes this issue by making sure that the cluster (the primary
and all remaining members) are correctly set with the SSL options
specified during the Cluster setup.

Change-Id: Ib66a18d109ff7f47e56877406c71c9dec2412868