BUG#35444244 rebootClusterFromCompleteOutage unable to rejoin instance after major outage
Rebooting from complete outage a Replica Cluster that can reach out to
its ClusterSet results in a failure rejoining the instances back to the
Cluster. The Cluster itself is rebooted and rejoined back to the
ClusterSet, however, the command fails to rejoin the instances.

This scenario is only reproducible when the Cluster is missing
transactions from the ClusterSet, or the Primary Cluster is under load.

The root cause is a missing transaction sync step after each instance's
recovery account is recreated. The account must be recreated on the
Primary Cluster so its replicated to the whole topology but since
there's no sync with the Replica Cluster, the account is still missing
when the instance is attempted to be rejoined.

This patch fixes that by making sure instances are only rejoined back to
their Clusters when the Cluster is rejoined back to the ClusterSet and
the transactions are sync'ed.

Change-Id: I58e1adcebe8907e77eccd17f81218afdfa9fddf5