Fixing build issue.
The shell code had a copy of the my_rnd function as it was not exposed in the server code.
Now the function is exposed on the server code and it caused a duplicate symbol issue.
This patch updates the shell code to use the function as available in the server.

Change-Id: If171ca5ca83e77d1b049917b0b0211119807e4b1