BUG#35528045 Exception executing Python command on Windows
Shell registers custom implementations of stdout and stderr for its
internal Python context. For those, it doesn't expose a fileno() method.
But in Windows, when a Python command is redirected (doesn't output to
the console), Python tries to access fileno(), and since it doesn't
exist, it throws an exception.

For the documentation, fileno() can simply throw an exception if it
doesn't have a valid file descriptor, as long as it is there. So this
patch exposes a fileno() method on both stdout and stderr, fixing the
underlying issue.

Change-Id: Idea4150fdb0c121a9a206b9bce8b3e5417805ecf