BUG#35468541 HEAD request is not retried if it fails with an authorization error
If an AWS HEAD request failed with an authorization error, it was not
retried, as response did not contain body, and retry mechanism uses body
to recognize AWS authorization errors.

As a workaround, whenever an AWS HEAD request fails with a 400 HTTP
error code, it is retried.

Additionally, the expiration time of AWS credentials (if present) is
adjusted, so that credentials are refreshed earlier, five minutes before
the required time.

Change-Id: I0a622383e01258284ac1f05e8d30f3f956d01e68
(cherry picked from commit d23d528ef370fbd450b7ec0666ccd4f67c4b2ce7)