BUG#35550282 Handle mysql_audit schema
Starting with 8.1.0 (8.0.34), MDS stores the audit plugin configuration
in the `mysql_audit` schema. This schema already exists when the MDS
instance is created, causing the dump&load from MDS into MDS to fail
with the duplicate objects error.

As a fix, when using `util.dumpInstance()` with the `ocimds` option set
to true, the `mysql_audit` schema is automatically excluded from the
dump.

Change-Id: I1e1329e4eb1ab395b9ddbf40fb2c9495c8b29e6f