BUG#35523549 BUG#35621465
WL#13688 introduced the "replicationSslMode" option when creating
ReplicaSets, to control the authentication type of the AR channels. It
didn't, however, expose the option on replicaset.options(). This patch
fixes this by exposing the option on that method.

Similarly, status() was also missing checking this option on the AR
channels, both on ReplicaSets and on the primary member of ClusterSets.
This patch fixes this issue by adding a new member, "replicationSslMode"
to the output of replicaset.status() and clusterset.status() (the latter
only when extended >= 1).

WL#13688 also introduced a bug, when creating ReplicaSets or
ClusterSets, where, if the user specified "memberAuthType" with the
value "certIssuer" or "certSubject", then the option
"clusterSetReplicationSslMode" (for ClusterSets) or the option
"replicationSslMode" (for ReplicaSets) would be ignored and always set
to "VERIFY_CA". This patch fixes this behavior, and the new option
"replicationSslMode" (aforementioned) confirms the fix.

Change-Id: I9f25d024626b62b0c299954eafe48dcd5fabb64e