BUG#35444206 When rebooting a cluster, the comm stack of the Cluster is ignored
When rebooting a Cluster that had its communication stack set to
"mysql", and where the user doesn't switch the stack, the primary would
be rebooted as if the stack was "xcom", not performing some checks or
even not re-creating recovery accounts, as required with using "mysql".

This patch fixes this by ensuring the the comm stack if read from the
Cluster's MD, if not specified, and used accordingly.

Change-Id: I84d8f3b652e5f20d5120359cb624faf38f511250