BUG#35434959 ReplicaSet.forcePrimaryInstance() fails not giving a reason
Currently, replicaSet.forcePrimaryInstance treats an instance that has
an applier state OFF, the same as if it was ERROR. This means that
the instance can't be used as a potential primary. But the fact that the
applier is stopped, doesn't mean that the instance can't be used as a
primary, only that we can't wait for transactions to be applied on it.

This patch changes the behavior of replicaSet.forcePrimaryInstance, so
that an instance with an applier with an OFF state is a valid candidate
for primary of the ReplicaSet. In this scenario, the
"invalidateErrorInstances" doesn't need to be specified (there's no
actual error) and the command doesn't wait for transactions to be
applied on that specific instance.

Change-Id: I35bd3d11c734f49a83638a03cdb75c9a86f6ad75