WL#15862 AdminAPI: Deprecations
The AdminAPI has several features/options that were deprecated several
releases ago, which were waiting for the next major release of MySQL to
be removed. With the transition into the new versioning/release model,
those can finally be removed in the next LTS version.

This WL adds the last deprecations before the removal.

In summary, the following features are deprecated:

  - Support for MySQL Server 5.7
  - dba.configureLocalInstance()
  - Cluster.checkInstanceState()

And the following options:

  - 'interactive' option in all AdminAPI
  - 'waitRecovery 'in favor of recoveryProgress
  - 'password' option in all AdminAPI
  - 'clearReadOnly' option in dba.configureInstance(),
    dba.configureLocalInstance(), and dba.dropMetadataSchema()

    * The option default is changed to 'true', i.e. super_read_only is
      automatically cleared when the option is not set.

Change-Id: Idab49463d460128faddc3677d6ec70d87d1ffda0