BUG#35594376 In a ClusterSet Error when Setting primary instance in Replica Cluster
Using Cluster.setPrimaryInstance() to elect a new primary on a Replica
Cluster but not actually changing the primary, i.e. the target instance
is already the primary, results in a broken ClusterSet replication
channel.

Since Group Replication doesn't allow changing the primary when there
are running replication channels at the target instance, the command
must stop the ClusterSet replication channel at the primary before
executing the switchover. The command's implementation does not check if
the target instance is already the current primary and proceeds with
that logic resulting in a stopped channel that must be restored later on
using ClusterSet.rejoinCluster()

This patch fixes that by checking first whether the target instance is
already the current primary to terminate the command right away
indicating that to the user.

Added regressions tests and also updated .cland-tidy to remove a
deprecation option: AnalyzeTemporaryDtors.

Change-Id: Ie9abb2e8b98559659e8582ad7b95e7f3fb6a7225