BUG#35449470 ClusterSet status reporting read replica instance as a cluster member
When a Replica Cluster with Read Replicas suffers a major outage,
ClusterSet.status({extended:1}) shows the Read Replicas as Cluster
members with the memberRole "SECONDARY", wrong status, and wrong
diagnostic information, instead of showing them as Read Replicas with
the right information.

The command is not differentiating a regular Cluster member from a Read
Replica in the scenario of missing Group Membership information
(outage), causing this issue.

This patch fixes that by adding the missing logic to handle this
particular scnenario in ClusterSet.status().

Change-Id: Idc37ad55d875aac24c6f813bf70962593ed07e4f