BUG#35391093 unable to obtain cluster handle from read-replica when quorum is lost
Attempting to obtain a cluster handle, using dba.getCluster(), while
connected to a Read Replica of a Cluster that has lost quorum results
in an error.

Quorum-loss should not make it impossible to get a cluster handle as
long as the active shell is established to a reachable Cluster member,
regardless if a Group Member or a Read Replica.

This patch fixes that by ensuring that the get_cluster() logic
automatically establishes a connection to a Cluster member to proceed
with the operation, instead of failing with an error indicating that no
primary member was found.

Change-Id: Ifebf16a76b7be2188478544bc304cbe0ee13ee97