WL#15887 BUG#35735901 BUG#35740174
WL#15887 Handle new SET_ANY_DEFINER privilege

New option
==========

The new option has been added to the following methods:

* `util.dumpInstance()`
* `util.dumpSchemas()`
* `util.dumpTables()`

targetVersion
-------------

string (default: current version of Shell) - Specifies version of the
destination MySQL server.

New behaviour
=============

The changes affect the following methods:

* `util.dumpInstance()`
* `util.dumpSchemas()`
* `util.dumpTables()`
* `util.copyInstance()`
* `util.copySchemas()`
* `util.copyTables()`

The MHS compatibility checks
----------------------------

 (`ocimds` option set to `true` or copy target is a DB System instance)

If the `targetVersion` options was implicitly set to the current version
of Shell, any compatibility error that would be printed for an older
version of MHS, is printed as a warning instead.

The following apply if target MHS version supports the `SET_ANY_DEFINER`
privilege (8.2.0+).

The statements with `DEFINER=` and `SQL SECURITY DEFINER` clauses no
longer result in a compatibility error. An error which requires manual
intervention is reported if the account specified as the definer has one
of the following user names:

* `mysql.infoschema`
* `mysql.session`
* `mysql.sys`
* `ociadmin`
* `ocidbm`
* `ocirpl`

A warning is reported if the account specified as the definer does not
exist or is excluded from the dump.

The `SET_ANY_DEFINER` privilege is added to the list of privileges
allowed in MHS.

The `strip_definers` compatibility option
-----------------------------------------

Using it when the `ocimds` option is enabled and the `targetVersion`
option is set to a version which supports the `SET_ANY_DEFINER`
privilege, results in a note stating that using this option is not
needed.

The `strip_restricted_grants` compatibility option
--------------------------------------------------

If the `strip_restricted_grants` compatibility option is used and the
`targetVersion` option is set to a version which supports the
`SET_ANY_DEFINER` privilege, the `SET_USER_ID` privilege is replaced
with `SET_ANY_DEFINER` instead of being stripped.

The copy utils
==============

The `targetVersion` option is internally set to the version of the
target instance.

BUG#35735901 Cosmetic: which is the latest version supported

The errors messages for unsupported versions of 'targetVersion' option
now include the minimum or maximum supported version.

BUG#35740174 "Incorrect" target version set implicitly

When using copy utilities with the 'strip_definers' compatibility
option set, the warning about this compatibility option and the
SET_ANY_DEFINER privilege was printed regardless of the version of the
target instance, this is now fixed.

Change-Id: Ic7cf6400f64c51dfe1132d5a009a0dcf161ca3ba
(cherry picked from commit c9bc985fe7fd6a0540f052149f16efcd14557dfb)