BUG#35548072 rebootClusterFromCompleteOutage fail to rejoin instanc if switch primary cluster
In case of a cluster switchover on a ClusterSet, where a replica Cluster
is OFFLINE (and which will be marked as invalid after the switchover),
trying to reboot that cluster will output error messages failing to
rejoin the cluster instances. The command will, however, continue and
the seed instance will be successfully rebooted. By coincidence, this
behavior is actually the correct one, because after a Cluster has been
invalidated, it must be manually rejoined to the ClusterSet and its
instances rejoined only after that.

This patch deals with this scenario properly, by not trying to rejoin
the instances and warning the user of the situation.

Change-Id: I2119c698aeda9dd2c5d9efc6723227627a934cb5