BUG#35600174 Memory consumption gradually increases when loader is using the subchunker
When loader loads a file which is bigger than `maxBytesPerTransaction`
(or 1.5 * `bytesPerChunk` if `maxBytesPerTransaction` was not used),
it is split into smaller chunks by the subchunker, storing these chunks
in memory. When this mode is active, file is being read into memory in
2MB chunks (actual size is controlled by the `net-buffer-length`
connection option). When subchunker was active, it was using .idx files
in order quickly find row boundaries, where subchunk would end. Due to
a bug, subchunker was always using the first unused entry in the .idx
file, instead of the one closest to the size of the allocated memory.
This meant that for every 2MB read, only ~1MB was consumed, so memory
utilization gradually increased.

As a fix, the .idx files are no longer used by the subchunker. Index
files contain entries for row boundaries, but they point to a first
boundary past every 1MB of data written, which means that with the
default `net-buffer-length` of 2MB, there is always going to be ~1MB of
data which remains in the buffer at all times.

Change-Id: I7392c08912fbe243e9e1e5559515ad7276423511