BUG#35635009 mysqlshell upgrade checker checkTableOutput stage mark view Corrupt
This commits fixes an issue with upgrade checker flagging views as corrupted in 5.7 (up to 39) when internal query contains a from statement with a table schema prefix with a "group by". The issue is only present, when we run the CHECK FOR UPGRADE query (that upgrade checker use) on a 5.7 (up to 39) MySQL database with no schema selected - so a safe "use mysql" statement was added for the local upgrade checker session to mitigate this problem.

Change-Id: I5c8cf09ffe207d5cd619dca52ef527f8b984bee8