WL#14870 BUG#33516540 BUG#33199178 BUG#35993066 BUG#35996854 BUG#35996514 BUG#35997891 BUG#35998172 BUG#36000967 AdminAPI: ReplicaSet rescan, dissolve and describe
This patch adds the following new commands to ReplicaSet:
 - dissolve() to dissolve the ReplicaSet
 - rescan() to update the ReplicaSet metadata, which includes updating
server ids, adding new instances (part of the topology but not present
in the MD, removing obsolete instances (not part of the topology but
still in the MD) and storing the current replication accounts.
 - describe() to output a description (JSON) of the ReplicaSet

It also deprecates the "addInstances" and "removeInstances" options from
Cluster.rescan() while adding two new ones, "addUnmanaged" and
"removeObsolete", replacing the deprecated ones with simpler options.

Regarding BUG#33516540, a new check was also added to status() to check
if the replication users aren't present in the MD, which might happen if
upgrading from a version older than 8.0.27. This causes unused
replication accounts to not be removed from the ReplicaSet after
removeInstance() removed the corresponding member. This new check in
status() informs the user to run rescan() which fixes the issue by
storing the account in the MD.

Regarding BUG#33199178, which occurs when calling setPrimaryInstance()
if the server_id var changed (e.g.: due an OS update), it's also fixed
by invoking rescan(), by storing the correct server_id in the MD.

Change-Id: Ia3ce3741ed2fbd8b3888475c51090b77614e7b7d