BUG#35590038 Attempt to load large file reports 100% completion 90 mins before actual complet
Once util.importTable() reaches 100% completion, it takes some time
before command actually completes. This is caused by the fact that
progress is reported based on the number of bytes read from the file,
while after the data is read, it takes some time for the server to
process the data.

In order to improve the user experience, an information about activity
of threads is now included in the progress information, i.e.:

8 thds loading - 77% (595.04 MB / 771.87 MB), 29.21 MB/s
6 thds loading - 2 thds committing - 89% (692.40 MB / 771.87 MB), 27.68 MB/s
2 thds committing - 100% (771.87 MB / 771.87 MB), 14.02 MB/s

Additionally, progress is now "animated", the first dash before
percentage is spinning.

Change-Id: I97bacd74b7126f621f72f9515d4fa38f182add6f