BUG#34969740 Failed dba.createCluster() may not rollback group_replication_*ssl* var changes
Commands in the AdminAPI should be, for the most part, transactional,
which means that, if the command aborts, it should leave the server or
servers as they were before the command was invoked. Currently, the
dba.createCluster() isn't reverting system variables back to their
original values.

This patch fixes this scenario and also when invoking
cluster.addInstance() and cluster.rejoinInstance().

Change-Id: I6ad1e5335718284c2935c9e8a2987705e2d2bbc8