Fix AdminAPI SSL check errors
WL15800 changed the default version and type of ciphers used in the
server. AdminAPI has some tests to check if an encrypted connection is
being used, but relies on a specific version and type of cipher.

This patch makes the check more generic, regardless of the type and
version of cipher used / chosen by the server.

Change-Id: I6ba26d4305ee81c1860ed0e2e1a66af32ae88396
(cherry picked from commit 18f0bd6caab62474c7bacbe1f3c2c0f4fed53196)