BUG#35370180 c.forceQuorumUsingPartitionOf() unnecessarily requires username
If the password and/or user for the admin account isn't specified
(either in the instance definition or explicitly by argument), the
command cluster.forceQuorumUsingPartitionOf() doesn't retrieve it from
current cluster session, like all other commands do. In regards to the
password, it will actually prompt the user for it.

This patch changes this behavior by retrieving the user and/or password
from the existing cluster session, if not explicitly supplied.

Change-Id: I2fca662c10b8787cb2a0f6c29a8d808609c7a6fb