BUG#33865464 Cluster.rescan() says I'm using X Protocol but I'm not
The error message being displayed is a generic one, where the shell
assumes that, if the connection fails, it's related to the X protocol.
This can lead to incorrect assumptions, specially if the connection
isn't using X protocol, times-out, and the error is shown talking about
the X protocol.

This patch solves this issue by simply throwing a generic error message,
not trying to find out if it's related to the X protocol, since that
check isn't foolproof.

Also cleaned the code to avoid leaking private declarations to entire
namespaces.

Also, since this issue was reported, BUG#35410360 changed how the X
port is checked in .status() and .rescan().

Change-Id: I1b1db0f1c1804935832ef725eaddbfea4deacbe3