BUG#35739776 Only one Read Replica is displayed when the label names are the same
When adding a read replica to a cluster, the label for that instance, if
specified, isn't validated, either its format or if an existing instance
is already using it.

This patch adds this verification, not allowing labels with invalid
characters to be used and by checking if the label specified is already
in use by another instance.

Change-Id: I32ac7965b6fd06e24a886eb7f4091352b8420039