BUG#35541522 Attempt to load large file into non-existing schema takes >20 min to fail
When importing a single, large, uncompressed file using
util.importTable() into a non-existing schema, users had to wait a
substantial amount of time for a failure. Additionally, with the fix for
BUG#35279351, operation would not terminate at all.

When a single file is imported, it is chunked into smaller parts. During
this operation errors reported by other threads were not checked, hence
operation failed only once the whole file was chunked.

As a fix, thread errors are checked frequently.

Change-Id: Iffe7f218c434a1af10ebc1053832115caadbdfed