BUG#35462985 util.exportTable() to remote location should not require a prefix to exist
Previously, trying to export a table to a path at remote location
required parent directory to exist, otherwise it failed with an error:

Cannot proceed with the dump, the directory containing
'table-export/emp.tsv' does not exist at the target location
table-export.

Given that cloud storage does not have a notion of directories (all
objects are files, which happen to have '/' characters in their name),
this requirement is superficial, and has been removed.

Change-Id: I7d8a96ca9ca2c3bc5a7c20a6080dcba82defb96f