BUG#35514497 double/wrong verification of group_replication_tls_source in internal checks
The AdminAPI needs to work with both new ("replica", "source") and old
("slave", "master") terminology. To accomplish this, it processes the
name of sysvars and replaces one keyword with another. The way it does
this however, isn't correct, because keywords like "replication" (which
shouldn't be change) are replaced with "slavetion".

This leads to a var like 'group_replication_tls_source' to be translated
to 'group_slavetion_tls_master', which lead to a double check (with the
incorrect name).

This patch addresses this issue by making sure that keywords are updated
only if delimited by either '_', '-', ' ', start or end of the name.

Change-Id: Iad530de3d0477bd49158f931a479039ab4c75218