Conditionally include FIDO-based client authentication plugins
FIDO-based client authentication plugins are excluded when build
uses a supported OpenSSL version (greather or equal than 1.1.1).

Additionally, the webauthn authentication plugin is only included
when using MySQL Server 8.2.0 where it was introduced.

This patch includes a new way to determine the MYSQL version as
the perl script used to determine it in windows no longer exists.

Change-Id: Iecfc339735326d00d808703a85e5d8ba6cab426c