Workaround for test-failure:
A bug in Group Replication can cause
dba.rebootClusterFromCompleteOutage() to hang when the MySQL
communication stack is used resulting in a test failure. The bug is
visible when a group_replication_start_on_boot is enabled.

The bug is not fixed yet, so this patch avoids it by disabling the
sysvar on this test which has no impact on the test purpose.

Change-Id: I9dd8d8db98c5f4ec92558e795a98e50991a3a88e