Bug#35782407 	cannot input full width digit 8
Full-width digit 8 is Unicode 0xFF18, which was being casted to char
when checked against a list of custom commands. Since the truncated 0x18
was a valid command keystroke, it would be intercepted instead of
inserted as part of input.

As a fix, characters above 255 are not checked against the list of
custom commands.

Change-Id: I82cb4e4008ff4a26c332596ccec519998756fd16