Added timeout for script test execution
Added --timeout option to run_unit_tests, which sets the number of
seconds to allow for each line in a test script to take. On timeout, the
trace output of the test is dumped and the entire test is aborted by
executing active interrupt handlers (without actually sending a ^C).
Default is 10 minutes, 0 disables the timeout.

Change-Id: I80d7ac4bec582b85df69b7b8c6884e6733af7062