Generalize topology executor
- Removed "cluster" from the topology executor name because it can also
be used with replica sets.
- Removed the requirement that the class op type must declare two
members of specific types ("m_undo_list" and "m_undo_tracker") to
support undo actions, even if it didn't use them.
- Forced the class op type to have a static function ("supports_undo")
to explicitly inform if it supports undo or not. If it does, then it
needs to support a non-static "do_undo" method, where the actual undo
action is performed.

Change-Id: I02fa336b3377c8a6f5caeeb14620c5c8d0647c22