BUG#35805866 Exclude mysql_firewall schema from dumps
Starting with 8.2.0 (8.0.35), MHS stores the firewall configuration
in the `mysql_firewall` schema. This schema already exists when the MHS
instance is created, causing the dump&load from MHS into MHS to fail
with the duplicate objects error.

As a fix, when using `util.dumpInstance()` with the `ocimds` option set
to true, the `mysql_firewall` schema is automatically excluded from the
dump.

Change-Id: I44fe0c6c3c5237ac35fa795b7f6c48b5e74e2f13