BUG#35356006 dba.rebootclusterfromcompleteoutage failed with default group_concat_max_len values
The dba.rebootclusterfromcompleteoutage needs to query the complete set
of GTIDs of a specific channel, and the SQL query that does that uses
GROUP_CONCAT function to aggregate results. However, the query can fail
if the limit set by the group_concat_max_len var is reached.

This patch created a version of the query that doesn't need to use the
GROUP_CONCAT function, which removes the issue completely. For scenarios
where it's still needed, the group_concat_max_len limit was raised to
1GB.

Change-Id: Ic097205d32fef4a5f86eeeada4c5acb5441a4720