BUG#35836245 BUG#35845603 Handling of removed GR options
BUG#35836245 Unnecessary handling of group_replication_view_change_uuid when server is 8.3+

The AdminAPI generates and sets the value for
group_replication_view_change_uuid when creating a Cluster with
dba.createCluster(), when the target instance has a version >= 8.0.27.
The value is stored in the Metadata and used when adding new instances
to the Cluster so that those instances use the same value.

This is needed to address the GTID-Set reconciliation in:

  - clusterset.forcePrimaryCluster()
  - clusterset.removeCluster()
  - replica_cluster.removeInstance()

The reconciliation is needed to avoid VCLEs not being able to replicate
when binglogs are purged.

In 8.3.0, Group Replication removes VCLEs from the binary logs and when
a Cluster is running 8.3.0+, group_replication_view_change_uuid is
ignored.

For that reason, the AdminAPI must not generate and set a value for
group_replication_view_change_uuid when a Cluster is created on a
target instance running 8.3.0+.

The instance on which the Cluster is created determines which will be
the lowest version in use by the Cluster, so we can safely assume
group_replication_view_change_uuid will never be needed.

This patch addresses that by ensuring the sysvar is only generated and
set when needed.

BUG#35845603 Unnecessary handling of group_replication_ip_whitelist when server is 8.3+

In 8.3.0, Group Replication removed support for
group_replication_ip_whitelist in favor of
group_replication_ip_allowlist.

However, the AdminAPI still list and accepts `ipWhitelist` as an option
for several commands, although it handles it automatically by
converting to the new terminology. But also, it still lists
group_replication_ip_whitelist in the output of Cluster.options().

This patch makes sure the sysvar is not queried and displayed in
cluster.options() when the target instance is running 8.3+.

Other changes:

  - Handling of binlog_transaction_dependency_tracking default value
    change to WRITESET
  - Updated dba.upgradeMetadata() help text to improve description of
    'dryRun' option

Change-Id: I0ff99735fb3be40f672961adfde754d4e10017e7