BUG#35845680 Unable to create Cluster due to removed option transaction_write_set_extraction
In 8.3.0, Group Replication removes the deprecated option
transaction_write_set_extraction (WL#13951). The option has been
deprecated since 8.0.26, but is in use by the AdminAPI to check and
enable parallel-appliers.

InnoDB Cluster dictates that transaction_write_set_extraction=XXHASH64
is a requirement, so the AdminAPI verifies the value set for it
in dba.checkInstanceConfiguration() (and implicitly when creating a
cluster with dba.createCluster()).

Also, it will set it when using dba.configureInstance().

With the removal of the option in 8.3.0, the commands above will
produce erroneous results and/or fail since the sysvar no longer
exists.

On top of that, Sandboxes are deployed with
transaction_write_set_extraction pre-configured too.

This patch fixes that by ensuring the commands no longer check and set
transaction_write_set_extraction.

Change-Id: I2d96f78a317ca717e213bee7a7051395fb57e616